gh-320 Rebased + refactoring of OAuthSecurityConfiguration
